+++ lib/controllers/me.ts
@@ -2,9 +2,9 @@ import { Express } from 'express'
 import Crowi from 'server/crowi'
 import Debug from 'debug'
 import fs from 'fs'
-import FileUploader from '../util/fileUploader'
-import GoogleAuth from '../util/googleAuth'
-import GitHubAuth from '../util/githubAuth'
+import FileUploader from '../utils/fileUploader'
+import GoogleAuth from '../utils/googleAuth'
+import GitHubAuth from '../utils/githubAuth'
 
 export default (crowi: Crowi, app: Express) => {
   const debug = Debug('crowi:routes:me')
