+++ lib/controllers/page.ts
@@ -1,8 +1,8 @@
 import { Types } from 'mongoose'
 import Crowi from 'server/crowi'
 import Debug from 'debug'
-import ApiResponse from '../util/apiResponse'
-import { decodeSpace } from '../util/path'
+import ApiResponse from '../utils/apiResponse'
+import { decodeSpace } from '../utils/path'
 import { BookmarkDocument } from 'server/models/bookmark'
 import { PageDocument } from 'server/models/page'
 import { RevisionDocument } from 'server/models/revision'
