+++ lib/controllers/revision.ts
@@ -1,6 +1,6 @@
 import Crowi from 'server/crowi'
 import Debug from 'debug'
-import ApiResponse from '../util/apiResponse'
+import ApiResponse from '../utils/apiResponse'
 
 export default (crowi: Crowi) => {
   const debug = Debug('crowi:routes:revision')
