+++ lib/middlewares/basicAuth.ts
@@ -1,7 +1,7 @@
 import { Express } from 'express'
 import Crowi from 'server/crowi'
 import basicAuth from 'basic-auth-connect'
-import { parseAccessToken } from '../util/accessTokenParser'
+import { parseAccessToken } from '../utils/accessTokenParser'
 
 export default (crowi: Crowi, app: Express) => {
   return (req, res, next) => {
