+++ lib/middlewares/swigFunctions.ts
@@ -1,6 +1,6 @@
 import { Express } from 'express'
 import Crowi from 'server/crowi'
-import functions from '../util/swigFunctions'
+import functions from '../utils/swigFunctions'
 
 export default (crowi: Crowi, app: Express) => {
   return (req, res, next) => {
