+++ lib/utils/linkDetector.ts
