+++ lib/utils/swigFunctions.ts
